requirements.txt
pydub
speechrecognition
python-telegram-bot==13.7
pydub==0.25.1
speechrecognition==3.8.1
pytube==11.0.1
ffmpeg-python==0.2.0